BackEnd/requirements.txt
flask>=2.3
flask-cors>=4.0
flask-sock>=0.7
python-dotenv>=1.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0

sqlalchemy
psycopg2
psycopg2-binary

numpy>=1.23
opencv-python>=4.8
librosa>=0.10
soundfile>=0.12
scikit-learn>=1.2
joblib>=1.2
soundfile>=0.12
python-dotenv
psycopg2-binary
google-generativeai
python-dotenv 
sqlalchemy 
psycopg2
reportlab>=4.0
